pyproject.toml
[project]
name = "agentic_code_editor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "groq>=1.1.1",
    "python-dotenv>=1.2.2",
    "rich>=14.3.3",
]
